add UART
